requirements.txt
scikit-learn
pandas
sentence-transformers
xgboost
nltk
transformers
faiss-cpu
torch